Removed many warnings
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@11005 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 024b97f6bff2e395953e0c40eb1d006a3d89e996

+++ application/SaLsA/de.offis.salsa.wrapper.car/src/de/uniol/inf/is/odysseus/salsa/wrapper/car/impl/CarSourceAdapter.java
@@ -159,7 +159,7 @@ public class CarSourceAdapter extends AbstractPushingSourceAdapter implements
 								calendar.add(Calendar.MILLISECOND,
 										millisecond * 10);
 								long timestamp = calendar.getTimeInMillis();
-								int id = (int) buffer.getShort();
+								int id = buffer.getShort();
 								int x = buffer.getInt();
 								int y = buffer.getInt();
 								double speed = Float.valueOf(buffer.getFloat())
